requirements.txt
aiohttp==3.8.5
aiomysql==0.2.0
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.4
dataclasses-json==0.5.14
docutils==0.16
fastapi==0.103.1
frozenlist==1.4.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-aiplatform==1.31.1
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.57.0
grpcio-status==1.57.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
langchain==0.0.292
langsmith==0.0.27
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
numexpr==2.8.5
numpy==1.25.2
openai==0.28.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.2
pip==23.2.1
platformdirs==3.10.0
proto-plus==1.22.3
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.12
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
rsa==4.7.2
s3transfer==0.6.2
setuptools==68.1.2
Shapely==1.8.5.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.4.0
tqdm==4.66.1
typing_extensions==4.7.1
typing-inspect==0.9.0
urllib3==1.26.16
uvicorn==0.23.2
wheel==0.41.2
yarl==1.9.2
